feat: cambiando a una arquitectura limpia (CA-007)
TODO: falta revisar que las dependencias sean las correctas, o sea, apunten a los módulos correctos para que conserven el principio de arquitecturas limpiar
