+++ core/entity/item.py
@@ -1,4 +1,4 @@
-from languageProcessing.nplModel import use_classifier
+from utils.languageProcessing.nplModel import use_classifier
 import os
 
 
@@ -37,7 +37,7 @@ class Item:
     @staticmethod
     def create_category(text):
         current_dir = os.getcwd()
-        path = os.path.join(current_dir, 'languageProcessing')
+        path = os.path.join(current_dir, 'utils', 'languageProcessing')
         category = text
         categoryNew = use_classifier(category, path)
         return categoryNew
