+++ core/usecase/test_downloader.py
@@ -1,6 +1,6 @@
 import unittest
-from crawler.request import Request
-from crawler.downloader import Downloader
+from core.entity.request import Request
+from core.usecase.downloader import Downloader
 
 
 class TestDownloader(unittest.TestCase):
