+++ infra/solr/config.py
