+++ infra/solr/solrManager.py
@@ -1,9 +1,9 @@
-from solr.solrManagerMeta import SolrManagerMeta
-from solr.solrCommands import (
+from infra.solr.solrManagerMeta import SolrManagerMeta
+from infra.solr.solrCommands import (
     CreateCoreCommand, CreateItemInCoreCommand,
     GetStatusCommand, ReloadCoreCommand
     )
-from solr.config import SOLR_BASE_URL
+from infra.solr.config import SOLR_BASE_URL
 import requests
