+++ infra/solr/solrManagerMeta.py
